**Please provide enough information so that others can review your pull request:**

Express supports regular expression with `?`, `+`, `*`, and `()` in their route path. So this PR adds the support for those too.
Unfortunately I have to use `regexp` since part of the route path will be a regular expression. If you have better suggestion than this, happy to make the change.

Note that even with the `regexp` introduced, it won't be used for all previous route path cases. This PR only uses it when the segment has a regular expression char (`?`, `+`, `*`, and `()`) in it, so the old code won't get the performance hit.

<!-- You can skip this if you're fixing a typo or adding an app to the Showcase. -->

**Explain the *details* for making this change. What existing problem does the pull request solve?**
Goal of this PR is to provide the same route functionality as Express does https://expressjs.com/en/guide/routing.html
Tests in this PR are also found in their document page.

<!-- Example: When "Adding a function to do X", explain why it is necessary to have a way to do X. -->

**Commit formatting** 

Use emojis on commit messages so it provides an easy way of identifying the purpose or intention of a commit. Check out the emoji cheatsheet here: https://gitmoji.carloscuesta.me/