@ReneWerner87 2020-07-25
Okay thanks the pull request, I will check the performance and if it does not affect the other cases, agree.  The logic of the given characters should even be quite easy to convert to string analysis without regex, maybe we will do that in the next step.

Hope the consumers of fiber don't get confused about the wildcard characters, because once it returns parameters and once it is in the part of a segment it doesn't.