@larrylv 2020-07-25
> The logic of the given characters should even be quite easy to convert to string analysis without regex
can you say more? It might require us building a subset of the regular expression engine, but I can't think of a way without regexp (not the package, but just regular expression) in general.

Here is the benchmark on master:
```
$ go test -v ./... -run=^$ -bench=Benchmark_Router_Github_API -benchmem -count=4
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber
Benchmark_Router_Github_API
Benchmark_Router_Github_API-12              1755            647651 ns/op             137 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1789            643144 ns/op             134 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1755            644639 ns/op             137 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1796            642922 ns/op             134 B/op          1 allocs/op
PASS
ok      github.com/gofiber/fiber        4.985s
PASS
ok      github.com/gofiber/fiber/middleware     0.113s
```

and here is with this PR
```
➜  ~/code/gopath/src/github.com/larrylv/fiber ruby:(2.7.0) go:(1.14.5) git:(larrylv/question_mark_in_route_path) ✔
$ go test -v ./... -run=^$ -bench=Benchmark_Router_Github_API -benchmem -count=4
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber
Benchmark_Router_Github_API
Benchmark_Router_Github_API-12              1664            707483 ns/op             181 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1614            686752 ns/op             187 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1718            696610 ns/op             176 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1729            680646 ns/op             175 B/op          1 allocs/op
PASS
ok      github.com/gofiber/fiber        5.131s
PASS
ok      github.com/gofiber/fiber/middleware     0.117s
```

The performance is degraded. I think it's probably because we will check if a Constant segment contains those regular expression characters https://github.com/gofiber/fiber/pull/673/commits/5d8cd24f0a4b46b25560e6d59f027070e4fe7f23#diff-167fa3abc5af405dd53a00a09a35b800R88

So I updated this PR a little bit in the last commit by only checking if the incremental Const part is regexp, and here is the _latest_ benchmark:
```
$ go test -v ./... -run=^$ -bench=Benchmark_Router_Github_API -benchmem -count=4
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber
Benchmark_Router_Github_API
Benchmark_Router_Github_API-12              1666            668311 ns/op             181 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1731            660890 ns/op             174 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1808            695656 ns/op             167 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1737            661348 ns/op             174 B/op          1 allocs/op
PASS
ok      github.com/gofiber/fiber        5.086s
PASS
ok      github.com/gofiber/fiber/middleware     0.205s
```

it is still degraded,  but that as much as before. Let me know if this is acceptable. I'll also think more about this to optimize this.