@ReneWerner87 2020-07-25
Okay, it looks pretty much the same as before, but today I don't have time 

We will see tomorrow if we can do this without affecting the normal routes 

Regarding the regex characters, this should also be possible with string analysis , or almost 90%....  The logics of the individual characters are well known, 

On Wildcard you just look until you find the substring after the characters, with plus the same, only that here at least one character must be present and question mark says only that the substring is optional, we have already implemented these logics in the parameters (wildcard, question mark)