@larrylv 2020-07-25
I would agree with you if we only allow those characters appearing once in one segment. Otherwise with the combination of various of them, it would be impossible to deal with only with string analysis. You would end up with a simplified regular expression engine.
The implementation of current wildcard and question mark is simple, yes, but it could only handle simpler cases:
1. if the whole segment is `*`
2. if the last char is `?`.
So it supports those two chars only in those scenarios.