@Fenny 2020-07-25
Hey @larrylv, thanks for the idea and the PR 👍 Regex has been a big topic the last months, @ReneWerner87 discussed to add support for it in `v1.12.0`. But we moved the idea because our current router is heavily optimized and we want to make sure it does not conflict with the current implementation.

That being said, I will make some time this week to dive deep into this PR/idea again regarding performance and compatibility. We also have to think about specific regex patterns that return param values.

To be continued....