@larrylv 2020-07-25
Thank you both for the quick review and feedback.
> We also have to think about specific regex patterns that return param values.

yeah, this PR only supports regexp with const segment, but doesn't support regexp with param. If that's a think gofiber also wants to support. I could give it a try after this PR.