@kiyonlin 2020-07-26
Hey @larrylv ! Thank you for your outstanding contribution.  For regular expression matching of routing paths, I have some ideas of my own.  

First of all, by using regular expressions, routing matching will become very flexible, but it will definitely bring performance problems.  Further more, attackers can construct very performance-consuming strings to attack(ddos) servers.  

Secondly, it is precisely because of flexibility that it will be difficult for us to extend route matching in the future—-no backward compatible.  

So, like @ReneWerner87 and @Fenny mentioned above, now we start with simple string parsing and slowly add more matching symbols and matching algorithms.  

In short, it is going to implement a simple regular expression matcher by ourselves.  If you are interested, you can participate in this work and improve it together!