@larrylv 2020-07-27
With the last two commits, I used the standard `strings` API `IndexAny` to find the index of the first appeared delimiter in `findNextRouteSegmentEnd`, and `ContainsAny` to check if it has any regexp character. Performance seems very close to master.
This PR:
```
$ go test -v ./... -run=^$ -bench=Benchmark_Router_Github_API -benchmem -count=4
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber
Benchmark_Router_Github_API
Benchmark_Router_Github_API-12              1665            669391 ns/op             181 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1677            676372 ns/op             180 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1718            668879 ns/op             176 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1766            663753 ns/op             171 B/op          1 allocs/op
PASS
ok      github.com/gofiber/fiber        4.995s
PASS
ok      github.com/gofiber/fiber/middleware     0.119s
```

master:
```
$ go test -v ./... -run=^$ -bench=Benchmark_Router_Github_API -benchmem -count=4
goos: darwin
goarch: amd64
pkg: github.com/gofiber/fiber
Benchmark_Router_Github_API
Benchmark_Router_Github_API-12              1617            669397 ns/op             149 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1712            668729 ns/op             140 B/op          1 allocs/op
Benchmark_Router_Github_API-12              1662            665205 ns/op             145 B/op          2 allocs/op
Benchmark_Router_Github_API-12              1686            661555 ns/op             143 B/op          1 allocs/op
PASS
ok      github.com/gofiber/fiber        4.988s
PASS
ok      github.com/gofiber/fiber/middleware     0.163s
```